// src/app/models/video-movie.interface.ts
// Generated by https://quicktype.io

export interface VideoListResponse {
    id: number;
    results: Video[];
}

export interface Video {
    iso_639_1: string;
    iso_3166_1: string;
    name: string;
    key: string;
    site: string;
    size: number;
    type: string;
    official: boolean;
    published_at: string;
    id: string;
}